Extract out lib
